requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
flask>=3.0.0
python-telegram-bot>=20.7

# Optional: For enhanced NLP capabilities
# nltk>=3.8.0
# spacy>=3.5.0

# Optional: For image processing
# Pillow>=10.0.0
# pytesseract>=0.3.10

# Optional: For search API integrations
# google-api-python-client>=2.100.0
# serpapi>=0.1.5

